Create PetrolPurchase
Create a class called PetrolPurchase to represent information about the petrol you purchase. The class should include five pieces of information in the form of instance variables-the station's location(type String), the type of petrol(type String), the quantity(type int), of the purchase in liter(double), and the percentage discount(double). Your class should have a constructor that initialize the five instance variables. Provide a set and get method for each instance variable. In addition, provide a method named getPurchaseAmount that caluclates the net purchase amount minus the discount, then return the net amount you had to pay as a double value. Write an application class named Petrol that demonstrates the capabilities of class PetrolPurchase