Car
Create a class called Car that includes three instance variables-a model(type String), a year(type String), and a price(type double). Provide a constructor and get and set methods for each initialized variable. The price must be positive. Apply a discount on both cars. 
Test the application with CarTest.